Fix 10 bugs from an adversarial codebase review
1. earnings.py used naive date.today() (server/UTC) instead of an
   SGT-aware date, misjudging same-day BMO/AMC releases for part of
   the SGT calendar day.
2. IBKR reconciliation truncated fractional share positions (DRIP,
   fractional-share brokers) to int, silently understating value/gain
   and re-flagging a spurious update every run. Shares now flow
   through as floats end to end (DB schema, /add, /update, IBKR
   parsing), with a shared format_shares() helper to keep whole-share
   display clean.
3. /alerts showed a near-support alert's distance from the
   auto-computed level, but the actual trigger uses a live price plus
   any manual ST/MT override — now resolved the same way in both
   places.
4. format_holdings_table() masked GAIN%/%CHG under privacy mode,
   contradicting /privacy's own documented behavior and how /list and
   market pings already treat those same percentages.
5. /week and /month counted deposits/withdrawals as investment
   performance. get_period_performance() now backs out the change in
   cost basis since the start snapshot.
6. Blocking Finnhub/yfinance/IBKR/FX calls ran directly on the bot's
   asyncio event loop, which it shares with Telegram long-polling and
   the scheduler — a slow fetch could stall both. Moved onto
   asyncio.to_thread across bot_handlers, telegram_handler, alerts,
   market_notifications, and ibkr_flex.
7 & 10. A Markdown-breaking symbol or error string reaching
   send_telegram_message() (e.g. IBKR-sourced data, which skips this
   bot's own symbol validation) could silently drop the whole message.
   It now retries once as plain text on a Telegram parse-error
   response.
8. /list, /watchlist, and /earnings could build a single message over
   Telegram's ~4096-char limit and get silently rejected. Added
   chunking (send_telegram_message and a bot_handlers _reply_chunked
   helper) that splits on line boundaries.
9. The IBKR shares/avg_cost comparison used exact equality; now that
   shares is a float, that could flap on float storage/round-trip
   noise. Both fields now use the same tolerance-based comparison.

+++ portfolio_db.py
@@ -17,12 +17,16 @@ def init_db():
     conn = _connect()
     cursor = conn.cursor()
     
-    # Holdings table
+    # Holdings table. shares is REAL, not INTEGER — fractional share counts
+    # are real (DRIP, fractional-share brokers, IBKR reconciliation). SQLite's
+    # dynamic typing already stored fractional values fine even under the old
+    # INTEGER declaration, so no migration is needed for existing databases —
+    # this only makes the declared type match what's actually stored.
     cursor.execute("""
         CREATE TABLE IF NOT EXISTS holdings (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             symbol TEXT NOT NULL UNIQUE,
-            shares INTEGER NOT NULL,
+            shares REAL NOT NULL,
             avg_cost REAL NOT NULL,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
